+++ bin/build.sh
@@ -19,5 +19,5 @@ touch docs/.nojekyll
 
 #clean up
 rm -rf src/*.js
-mv dist/some-sql.min.js dist/some-sql.min.0.1.4.js
-echo "$(cat dist/some-sql.min.0.1.4.js)" | gzip -9f | wc -c;
+mv dist/some-sql.min.js dist/some-sql.min.0.1.5.js
+echo "$(cat dist/some-sql.min.0.1.5.js)" | gzip -9f | wc -c;