+++ memory-db.d.ts
@@ -1,4 +1,4 @@
-import { SomeSQLBackend, ActionOrView, QueryLine } from "./index";
+import { SomeSQLBackend, ActionOrView, QueryLine, DataModel } from "./index";
 import { TSMap } from "typescript-map";
 /**
  * In memory storage implimentation.
@@ -12,7 +12,7 @@ export declare class SomeSQLMemDB implements SomeSQLBackend {
     /**
      * Creates all the tables and prepares the database for use.
      *
-     * @param {TSMap<string, Array<Object>>} models
+     * @param {TSMap<string, Array<DataModel>>} models
      * @param {TSMap<string, Array<ActionOrView>>} actions
      * @param {TSMap<string, Array<ActionOrView>>} views
      * @param {TSMap<string, Function>} filters
@@ -21,7 +21,7 @@ export declare class SomeSQLMemDB implements SomeSQLBackend {
      *
      * @memberOf SomeSQLMemDB
      */
-    connect(models: TSMap<string, Array<Object>>, actions: TSMap<string, Array<ActionOrView>>, views: TSMap<string, Array<ActionOrView>>, filters: TSMap<string, Function>, preCustom: Array<any>, callback: Function): void;
+    connect(models: TSMap<string, Array<DataModel>>, actions: TSMap<string, Array<ActionOrView>>, views: TSMap<string, Array<ActionOrView>>, filters: TSMap<string, Function>, preCustom: Array<any>, callback: Function): void;
     /**
      * Public exec option.  Organizes the query then sends it to the internal execution function.
      *